Add files via upload
Creating a Home WIFI Network using cisco packet tracer.8.0.2. This is just a simple and basic demonstration for those interested in a step on how home WIFI network is connected. Through 3 laptops and a home router wirelessly. this demonstration is shown in a simple logical view. In addition it there is a step by step guide which can help beginners on how to get started with cisco packet tracer.8.0.2 with a simple documentation that was made with some instructions.